Moving utility function to service
